pyproject.toml
[project]
name = "ghgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paolo Tranquilli", email = "redsun82@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "inflection>=0.5.1",
    "pystache>=0.6.8",
    "pytest>=8.3.4",
    "ruamel-yaml>=0.18.10",
]

[project.scripts]
gh-gen = "ghgen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
]
